fix(APISIMP-ADMIN-*-NO-PAGINATION): fix integer overflow in page*pageSize (CodeQL CWE-190)
Cast `page * pageSize` to `long` before multiplication in all 4 admin list
endpoints to prevent int overflow when page is large. Also cast `from + pageSize`
inside Math.min via long arithmetic. Fixes 8 CodeQL HIGH alerts in PR #2563.

+++ backend/src/main/java/de/dlr/shepard/v2/admin/plugins/PluginsAdminRest.java
@@ -141,10 +141,10 @@ public class PluginsAdminRest {
     for (PluginEntry entry : entries) {
       rows.add(PluginEntryIO.from(entry, registry.isEnabled(entry.id())));
     }
-    int from = page * pageSize;
+    long from = (long) page * pageSize;
     List<PluginEntryIO> slice = from >= rows.size()
         ? List.of()
-        : rows.subList(from, Math.min(from + pageSize, rows.size()));
+        : rows.subList((int) from, (int) Math.min(from + pageSize, rows.size()));
     return Response.ok(new PagedResponseIO<>(slice, rows.size(), page, pageSize))
         .build();
   }
